+++ lib/src/templates/widgets/appbar_template.dart
@@ -1,6 +1,7 @@
 import 'package:curie/src/views/settings/settings_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:go_router/go_router.dart';
 
 AppBar appBarTemplate({
   required BuildContext context,
@@ -27,12 +28,7 @@ AppBar appBarTemplate({
           // to the app after it has been killed while running in the
           // background, the navigation stack is restored.
           // Navigator.restorablePushNamed(context, SettingsView.routeName); // TODO remove line after using go_router
-          Navigator.push(
-            context,
-            MaterialPageRoute(
-              builder: (BuildContext context) => const SettingsView(),
-            ),
-          );
+          context.push(SettingsView.routeName);
         },
       ),
     ],