+++ lib/src/views/error/error_view.dart
@@ -1,21 +1,33 @@
 import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
 
+/// Default error page implementation for Material.
 class ErrorView extends StatelessWidget {
   const ErrorView({
-    required this.errorDetails,
+    this.errorDetails,
+    this.error,
     Key? key,
-  }) : super(key: key);
+  })  : assert(
+          errorDetails != null || error != null,
+          'At least one parameter must be non-null',
+        ),
+        super(key: key);
 
-  final FlutterErrorDetails errorDetails;
+  /// The FlutterErrorDetails to be displayed.
+  final FlutterErrorDetails? errorDetails;
+
+  /// The exception to be displayed.
+  final Exception? error;
 
   @override
   Widget build(BuildContext context) {
+    // TODO translate
     return SafeArea(
       child: Scaffold(
         appBar: AppBar(
           backgroundColor: Theme.of(context).colorScheme.error,
           title: Text(
-            'Erro',
+            'Error',
             style: TextStyle(
               fontSize: 24.0,
               fontWeight: FontWeight.w400,
@@ -28,14 +40,23 @@ class ErrorView extends StatelessWidget {
             mainAxisSize: MainAxisSize.min,
             children: [
               Text(
-                'An unknown error has occurred.',
+                errorDetails == null
+                    ? 'Page not found'
+                    : 'An unknown error has occurred.',
                 textAlign: TextAlign.center,
                 style: Theme.of(context).textTheme.displayMedium,
               ),
               const SizedBox(height: 20),
-              Text(
-                errorDetails.exceptionAsString(),
+              SelectableText(
+                errorDetails == null
+                    ? error?.toString() ?? 'Page not found'
+                    : errorDetails!.exceptionAsString(),
                 style: Theme.of(context).textTheme.bodyLarge,
+              ),
+              const SizedBox(height: 20),
+              TextButton(
+                onPressed: () => context.go('/'),
+                child: const Text('Home'),
               ),
             ],
           ),