+++ lib/src/views/route_test/route_test_view.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class RouteTestView extends StatelessWidget {
+  const RouteTestView({super.key});
+
+  static const routeName = '/test/:page';
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}